+++ src/renderer/src/pages/QuizSession/QuizSessionPage.tsx
@@ -1,4 +1,4 @@
-import { Box, Button, Card, Group, Radio, Stack, Textarea, Title } from '@mantine/core'
+import { Box, Button, Card, Group, Radio, Stack, Textarea } from '@mantine/core'
 import { useForm } from '@mantine/form'
 import { Question, Quiz, submissionSchema } from '@noggin/types/quiz-types'
 import { useModule } from '@renderer/app/hooks/use-module'